requirements.txt
streamlit==1.46.1
google-generativeai==0.8.5
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0 